requirements.txt
# STGNN项目依赖（已移除RL相关包）

# 核心依赖
numpy>=1.21.0
torch>=1.12.0

# 图处理（STGNN需要）
networkx>=2.8
torch-geometric>=2.0.0  # 用于GNN

# 数据处理
pandas>=1.3.0
scipy>=1.7.0

# 地理空间数据处理
geopandas>=0.12.0
shapely>=2.0.0
ezdxf>=1.0.0  # DXF文件导入

# 配置管理
pyyaml>=6.0

# 可视化
matplotlib>=3.5.0
seaborn>=0.12.0  # 可选，用于高级可视化

# 超参数搜索
optuna>=3.0.0

# 测试
pytest>=7.0.0